[refactor] 사용자 초기 데이터 위치 변경
